+++ src/vanillaExtract.ts
@@ -28,7 +28,14 @@ type ClassNameCallback = (
 ) => string;
 
 /**
- * Returns a scoped cssClassName styled with free-form CSS
+ * Returns a scoped cssClassName styled with free-form CSS. This function is a
+ * thin wrapper around vanilla-extract's `style` function.
+ *
+ * When you pass it a string, all `.&` tokens are automatically replaced with the
+ * selector for the auto-generated class-name. Note that in such cases EVERY
+ * style property must be wrapped in a selector block.
+ *
+ * To opt out of the `.&` replacement, use the callback function signature.
  *
  * @see https://github.com/reykjavikcity/webtools/blob/v0.1/README.md#vanillaclass
  */
@@ -41,12 +48,20 @@ export function vanillaClass(
 ): string {
   const debugId = css != null ? (cssOrDebugId as string) : undefined;
   css = css != null ? css : cssOrDebugId;
-  if (typeof css === 'function') {
-    const className = style({}, debugId);
-    vanillaGlobal(css('className', `.${className}`));
-    return className;
+
+  if (typeof css === 'string' && !/.&/.test(css)) {
+    return style(vanillaProps(css), debugId);
   }
-  return style(vanillaProps(css), debugId);
+
+  const className = style({}, debugId);
+
+  vanillaGlobal(
+    typeof css === 'function'
+      ? css('className', `.${className}`)
+      : css.replace(/.&/g, `.${className}`)
+  );
+
+  return className;
 }
 
 // ---------------------------------------------------------------------------